refactor: moved and renamed stuff into something easier to understand
