+++ fhir-test-services/fhir-itb-service/src/main/java/eu/europa/ec/fhir/handlers/HandlersConfig.java
@@ -1,5 +1,6 @@
 package eu.europa.ec.fhir.handlers;
 
+import eu.europa.ec.fhir.proxy.ItbRestClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
