+++ fhir-test-services/fhir-itb-service/src/main/java/eu/europa/ec/fhir/http/Response.java
@@ -1,16 +1,16 @@
-package eu.europa.ec.fhir.handlers;
+package eu.europa.ec.fhir.http;
 
 import java.net.http.HttpHeaders;
 import java.util.Optional;
 
 /**
- * Record reflecting a request's result.
+ * An HTTP response data.
  *
  * @param status  The HTTP status code.
  * @param body    The response's body.
  * @param headers The returned headers.
  */
-public record RequestResult(int status, String body, HttpHeaders headers) {
+public record Response(int status, String body, HttpHeaders headers) {
 
     public Optional<String> contentType() {
         return headers.allValues(org.springframework.http.HttpHeaders.CONTENT_TYPE)
