+++ fhir-test-services/fhir-itb-service/src/main/java/eu/europa/ec/fhir/state/SavedPost.java
@@ -1,6 +1,6 @@
 package eu.europa.ec.fhir.state;
 
-import eu.europa.ec.fhir.handlers.RequestResult;
+import eu.europa.ec.fhir.http.Response;
 
 /**
  * Information on a received POST for a patient that was not immediately matched to a test session.
@@ -9,5 +9,5 @@ import eu.europa.ec.fhir.handlers.RequestResult;
  * @param body The POST body.
  * @param serverResult The result returned to the caller (produced from the embedded FHIR server).
  */
-public record SavedPost(String patient, String body, RequestResult serverResult) {
+public record SavedPost(String patient, String body, Response serverResult) {
 }
